wip: add repo to each database model
