+++ web-app/src/app/api/stripe/webhook/route.ts
@@ -2,12 +2,12 @@ import { NextResponse } from "next/server";
 import Stripe from "stripe";
 
 import { getEnvSecrets } from "@/config/env.config";
-import { prisma } from "@/lib/db";
 import {
   getFiatTransactionByServicePaymentId,
+  prisma,
   setFiatTransactionStatusToFailed,
   setFiatTransactionStatusToSucceeded,
-} from "@/lib/services/fiatTransaction.service";
+} from "@/lib/db";
 import { FiatTransactionStatus } from "@/prisma/generated/client";
 
 const stripe = new Stripe(getEnvSecrets().STRIPE_SECRET_KEY);
