+++ web-app/src/app/app/agents/page.tsx
@@ -2,8 +2,7 @@ import { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
 
 import { requireAuthentication } from "@/lib/auth/utils";
-import { AgentWithRelations } from "@/lib/db";
-import { getAgents } from "@/lib/services/agent.service";
+import { AgentWithRelations, getAgents } from "@/lib/db";
 import { getOrCreateFavoriteAgentList } from "@/lib/services/agentList.service";
 import { getAgentCreditsPrice } from "@/lib/services/credit.service";
 import { getTags } from "@/lib/services/tag.service";
