+++ web-app/src/lib/db/agentList/index.ts
@@ -1,1 +1,2 @@
+export * from "./repo";
 export * from "./types";
