+++ web-app/src/lib/db/credit/index.ts
@@ -1,2 +1,3 @@
+export * from "./repo";
 export * from "./types";
 export * from "./utils";
