+++ web-app/src/lib/db/creditCost/index.ts
@@ -1,3 +1,2 @@
 export * from "./repo";
-export * from "./types";
 export * from "./utils";
