+++ web-app/src/lib/db/creditCost/utils.ts
@@ -1,0 +1,3 @@
+export function parseUnit(unit: string) {
+  return unit === "lovelace" ? "" : unit;
+}
