+++ web-app/src/lib/db/fiatTransaction/index.ts
@@ -1,0 +1,1 @@
+export * from "./repo";
