+++ web-app/src/lib/db/index.ts
@@ -1,5 +1,7 @@
 export * from "./agent";
 export * from "./agentList";
 export * from "./credit";
+export * from "./creditCost";
+export * from "./fiatTransaction";
 export * from "./job";
 export { default as prisma } from "./prisma";
