+++ web-app/src/lib/services/agentList.service.ts
@@ -1,63 +1,13 @@
 "use server";
 
-import { agentListInclude, AgentListWithAgent, prisma } from "@/lib/db";
-import { AgentList, AgentListType, Prisma } from "@/prisma/generated/client";
-
-export async function getAgentLists(
-  userId: string,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<AgentListWithAgent[]> {
-  const agentLists = await tx.agentList.findMany({
-    where: { userId },
-    include: agentListInclude,
-    orderBy: {
-      type: "asc",
-    },
-  });
-
-  if (!agentLists) {
-    return [];
-  }
-
-  return agentLists;
-}
-
-export async function getAgentListByType(
-  userId: string,
-  type: AgentListType,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<AgentListWithAgent | null> {
-  const agentList = await tx.agentList.findFirst({
-    where: {
-      userId,
-      type,
-    },
-    include: agentListInclude,
-  });
-
-  return agentList;
-}
-
-export async function createAgentList(
-  userId: string,
-  type: AgentListType,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<AgentListWithAgent> {
-  return await tx.agentList.create({
-    data: {
-      userId,
-      type,
-    },
-    include: agentListInclude,
-  });
-}
-
-export async function getOrCreateFavoriteAgentList(
-  userId: string,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<AgentListWithAgent> {
-  return await getOrCreateAgentListByType(userId, AgentListType.FAVORITE, tx);
-}
+import {
+  AgentListWithAgent,
+  createAgentList,
+  getAgentListByType,
+  getAgentListsByTypes,
+  prisma,
+} from "@/lib/db";
+import { AgentListType, Prisma } from "@/prisma/generated/client";
 
 export async function getOrCreateAgentListByType(
   userId: string,
@@ -73,21 +23,28 @@ export async function getOrCreateAgentListByType(
   return await createAgentList(userId, type, tx);
 }
 
+export async function getOrCreateFavoriteAgentList(
+  userId: string,
+  tx: Prisma.TransactionClient = prisma,
+): Promise<AgentListWithAgent> {
+  return await getOrCreateAgentListByType(userId, AgentListType.FAVORITE, tx);
+}
+
+export async function getFavoriteAgents(
+  userId: string,
+  tx: Prisma.TransactionClient = prisma,
+) {
+  const list = await getOrCreateFavoriteAgentList(userId, tx);
+  return list.agents;
+}
+
 export async function getOrCreateAgentListsByTypes(
   userId: string,
   types: AgentListType[],
   tx: Prisma.TransactionClient = prisma,
 ): Promise<AgentListWithAgent[]> {
   // Get all existing lists for the user that match the requested types
-  const existingLists = await tx.agentList.findMany({
-    where: {
-      userId,
-      type: {
-        in: types,
-      },
-    },
-    include: agentListInclude,
-  });
+  const existingLists = await getAgentListsByTypes(userId, types, tx);
 
   // Find which types are missing
   const existingTypes = new Set(existingLists.map((list) => list.type));
@@ -95,45 +52,11 @@ export async function getOrCreateAgentListsByTypes(
 
   // Create missing lists
   const newLists = await Promise.all(
-    missingTypes.map((type) =>
-      tx.agentList.create({
-        data: {
-          userId,
-          type,
-        },
-        include: agentListInclude,
-      }),
-    ),
+    missingTypes.map((type) => createAgentList(userId, type, tx)),
   );
 
   // Combine existing and new lists, sorted by type
   return [...existingLists, ...newLists].sort((a, b) =>
     a.type.localeCompare(b.type),
   );
-}
-
-export async function addAgentToList(
-  agentId: string,
-  listId: string,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<AgentList> {
-  return await tx.agentList.update({
-    where: { id: listId },
-    data: {
-      agents: { connect: { id: agentId } },
-    },
-  });
-}
-
-export async function removeAgentFromList(
-  agentId: string,
-  listId: string,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<AgentList> {
-  return await tx.agentList.update({
-    where: { id: listId },
-    data: {
-      agents: { disconnect: { id: agentId } },
-    },
-  });
 }
