+++ web-app/src/lib/services/fiatTransaction.service.ts
@@ -4,81 +4,13 @@ import {
   createCheckoutSession,
   getConversionFactors,
 } from "@/lib/actions/stripe.actions";
-import { prisma } from "@/lib/db";
 import {
-  FiatTransaction,
-  FiatTransactionStatus,
-  Prisma,
-} from "@/prisma/generated/client";
+  createFiatTransaction,
+  prisma,
+  updateFiatTransactionServicePaymentId,
+} from "@/lib/db";
 
 import { getUserById } from "./user.service";
-
-async function createFiatTransaction(
-  userId: string,
-  cents: bigint,
-  centsPerAmount: bigint,
-  currency: string,
-  tx: Prisma.TransactionClient = prisma,
-) {
-  const fiatTransaction = await tx.fiatTransaction.create({
-    data: {
-      userId,
-      centsPerAmount,
-      amount: cents / centsPerAmount,
-      currency,
-    },
-  });
-  return fiatTransaction;
-}
-
-export async function getFiatTransactionByServicePaymentId(
-  servicePaymentId: string,
-  tx: Prisma.TransactionClient = prisma,
-) {
-  return await tx.fiatTransaction.findUnique({
-    where: { servicePaymentId },
-  });
-}
-
-export async function updateServicePaymentId(
-  fiatTransactionId: string,
-  servicePaymentId: string,
-  tx: Prisma.TransactionClient = prisma,
-) {
-  const fiatTransaction = await tx.fiatTransaction.update({
-    where: { id: fiatTransactionId },
-    data: { servicePaymentId },
-  });
-  return fiatTransaction;
-}
-
-export async function setFiatTransactionStatusToSucceeded(
-  fiatTransaction: FiatTransaction,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<FiatTransaction> {
-  return await tx.fiatTransaction.update({
-    where: { id: fiatTransaction.id },
-    data: {
-      status: FiatTransactionStatus.SUCCEEDED,
-      creditTransaction: {
-        create: {
-          userId: fiatTransaction.userId,
-          amount: fiatTransaction.amount * fiatTransaction.centsPerAmount,
-        },
-      },
-    },
-  });
-}
-
-export async function setFiatTransactionStatusToFailed(
-  fiatTransaction: FiatTransaction,
-  tx: Prisma.TransactionClient = prisma,
-): Promise<FiatTransaction> {
-  return await tx.fiatTransaction.update({
-    where: { id: fiatTransaction.id },
-    data: { status: FiatTransactionStatus.FAILED },
-  });
-}
 
 export async function createStripeCheckoutSession(
   userId: string,
@@ -105,7 +37,11 @@ export async function createStripeCheckoutSession(
       Number(fiatTransaction.amount) /
         conversionFactorsPerCredit.amountPerCredit,
     );
-    await updateServicePaymentId(fiatTransaction.id, stripeSessionId, tx);
+    await updateFiatTransactionServicePaymentId(
+      fiatTransaction.id,
+      stripeSessionId,
+      tx,
+    );
     return { stripeSessionId, url };
   });
 }
